+++ Tests/EntityFrameworkCore.CosmosDb.TestApplication/EntityFrameworkCore.CosmosDb.TestApplication.Domain/Entities/Associations/S_NoPkInCompositeDependent.cs
@@ -1,0 +1,11 @@
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: IntentTemplate("Intent.Entities.DomainEntity", Version = "2.0")]
+
+namespace EntityFrameworkCore.CosmosDb.TestApplication.Domain.Entities.Associations
+{
+    public class S_NoPkInCompositeDependent
+    {
+        public string Description { get; set; }
+    }
+}