+++ Tests/EntityFrameworkCore.CosmosDb.TestApplication/EntityFrameworkCore.CosmosDb.TestApplication.Domain/Repositories/Associations/IS_NoPkInCompositeRepository.cs
@@ -1,0 +1,21 @@
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using EntityFrameworkCore.CosmosDb.TestApplication.Domain.Entities.Associations;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Entities.Repositories.Api.EntityRepositoryInterface", Version = "1.0")]
+
+namespace EntityFrameworkCore.CosmosDb.TestApplication.Domain.Repositories.Associations
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    public interface IS_NoPkInCompositeRepository : IEFRepository<S_NoPkInComposite, S_NoPkInComposite>
+    {
+        [IntentManaged(Mode.Fully)]
+        Task<S_NoPkInComposite?> FindByIdAsync(Guid id, CancellationToken cancellationToken = default);
+        [IntentManaged(Mode.Fully)]
+        Task<List<S_NoPkInComposite>> FindByIdsAsync(Guid[] ids, CancellationToken cancellationToken = default);
+    }
+}