+++ Tests/EntityFrameworkCore.CosmosDb.TestApplication/EntityFrameworkCore.CosmosDb.TestApplication.Infrastructure/DependencyInjection.cs
@@ -73,6 +73,8 @@ namespace EntityFrameworkCore.CosmosDb.TestApplication.Infrastructure
             services.AddTransient<IP_SourceNameDiffRepository, P_SourceNameDiffRepository>();
             services.AddTransient<IQ_DestNameDiffRepository, Q_DestNameDiffRepository>();
             services.AddTransient<IR_SourceNameDiffRepository, R_SourceNameDiffRepository>();
+            services.AddTransient<IS_NoPkInCompositeRepository, S_NoPkInCompositeRepository>();
+            services.AddTransient<IT_NoPkInCompositeRepository, T_NoPkInCompositeRepository>();
             services.AddTransient<IAudit_DerivedClassRepository, Audit_DerivedClassRepository>();
             services.AddTransient<IAudit_SoloClassRepository, Audit_SoloClassRepository>();
             services.AddTransient<IFolderEntityRepository, FolderEntityRepository>();