+++ Tests/EntityFrameworkCore.CosmosDb.TestApplication/EntityFrameworkCore.CosmosDb.TestApplication.Infrastructure/Persistence/Configurations/Associations/E_RequiredCompositeNavConfiguration.cs
@@ -32,8 +32,7 @@ namespace EntityFrameworkCore.CosmosDb.TestApplication.Infrastructure.Persistenc
 
         public void ConfigureE_RequiredDependent(OwnedNavigationBuilder<E_RequiredCompositeNav, E_RequiredDependent> builder)
         {
-            builder.WithOwner(x => x.E_RequiredCompositeNav)
-                .HasForeignKey(x => x.Id);
+            builder.WithOwner(x => x.E_RequiredCompositeNav);
 
             builder.HasKey(x => x.Id);