+++ Tests/EntityFrameworkCore.CosmosDb.TestApplication/EntityFrameworkCore.CosmosDb.TestApplication.Infrastructure/Repositories/Associations/T_NoPkInCompositeRepository.cs
@@ -1,0 +1,37 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using AutoMapper;
+using EntityFrameworkCore.CosmosDb.TestApplication.Domain.Entities.Associations;
+using EntityFrameworkCore.CosmosDb.TestApplication.Domain.Repositories;
+using EntityFrameworkCore.CosmosDb.TestApplication.Domain.Repositories.Associations;
+using EntityFrameworkCore.CosmosDb.TestApplication.Infrastructure.Persistence;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.EntityFrameworkCore.Repositories.Repository", Version = "1.0")]
+
+namespace EntityFrameworkCore.CosmosDb.TestApplication.Infrastructure.Repositories.Associations
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    public class T_NoPkInCompositeRepository : RepositoryBase<T_NoPkInComposite, T_NoPkInComposite, ApplicationDbContext>, IT_NoPkInCompositeRepository
+    {
+        public T_NoPkInCompositeRepository(ApplicationDbContext dbContext, IMapper mapper) : base(dbContext, mapper)
+        {
+        }
+
+        public async Task<T_NoPkInComposite?> FindByIdAsync(Guid id, CancellationToken cancellationToken = default)
+        {
+            return await FindAsync(x => x.Id == id, cancellationToken);
+        }
+
+        public async Task<List<T_NoPkInComposite>> FindByIdsAsync(
+            Guid[] ids,
+            CancellationToken cancellationToken = default)
+        {
+            return await FindAllAsync(x => ids.Contains(x.Id), cancellationToken);
+        }
+    }
+}