+++ Modules/Intent.Modules.EntityFrameworkCore/release-notes.md
@@ -1,6 +1,7 @@
 ﻿### Version 4.6.1
 
 - Improvement: Added support for Oracle databases.
+- Fixed: Using CosmosDB on a one-to-one relationship will no longer generate an implicit PK field on the composite entity.
 
 ### Version 4.6.0